Added text color animation and padding for the Label.
